Fix Champ_front_recyclage
We got an error when using the keyword 'moyenne_imposee connexion_exacte' when all procs intented to read in the same file at the same time.
We now use LecFicDiffuse instead of EFichier.

We added the keyword in the Test_recycl_lecture test case, and changed
the reference because of course we changed the result by adding a new
boundary condition.