initialize nbc variable for KEps_Meshing_VEF conversion + f-string
The nbc variable was not initialized and caused an error in the conversion of the KEps_Meshing_VEF
prm form in TrioCFD. My editor uses ruff to impose pycodestyle. I also use a linter that points
unused variable or functions. I converted strings that uses the percent style to the modern f-string
for readability.